Check empty for POST variable

+++ plugins/githuber-enhanced-plugin/inc/class-post-type-repository.php
@@ -70,7 +70,7 @@ class Post_Type_Repository {
 		add_meta_box(
 			'repository_meta_box',            // id.
 			'GitHub Repository',              // title.
-			array( $this, 'show_meta_box' ), // callback.
+			array( $this, 'show_meta_box' ),  // callback.
 			'repository',                     // screen.
 			'normal',                         // context.
 			'high'                            // priority.
@@ -140,13 +140,15 @@ class Post_Type_Repository {
 			}
 		}
 
-		$old = get_post_meta( $post_id, 'github_repository', true );
-		$new = $_POST['github_repository'];
-
-		if ( $new && $new !== $old ) {
-			update_post_meta( $post_id, 'github_repository', $new );
-		} elseif ( '' === $new && $old ) {
-			delete_post_meta( $post_id, 'github_repository', $old );
+		if ( !empty( $_POST['github_repository'] ) ) {
+			$old = get_post_meta( $post_id, 'github_repository', true );
+			$new = $_POST['github_repository'];
+	
+			if ( $new && $new !== $old ) {
+				update_post_meta( $post_id, 'github_repository', $new );
+			} elseif ( '' === $new && $old ) {
+				delete_post_meta( $post_id, 'github_repository', $old );
+			}
 		}
 	}
